+++ b_cfn_custom_api_key_authorizer_layer/authorizer_layer.py
@@ -26,6 +26,5 @@ class AuthorizerLayer(LayerVersion):
     @staticmethod
     def runtimes() -> Optional[List[Runtime]]:
         return [
-            Runtime.PYTHON_3_8,
-            Runtime.PYTHON_3_9
+            Runtime.PYTHON_3_8
         ]