+++ b_cfn_custom_api_key_authorizer_test/integration/infrastructure/main_stack.py
@@ -1,13 +1,17 @@
-from aws_cdk.aws_apigatewayv2 import CfnApi, CfnStage
 from aws_cdk.core import Construct
 from b_aws_testing_framework.tools.cdk_testing.testing_stack import TestingStack
 
-from b_cfn_custom_api_key_authorizer.custom_authorizer import ApiKeyCustomAuthorizer
-from b_cfn_custom_api_key_authorizer_test.integration.infrastructure import authorized_endpoint_stack
+from b_cfn_custom_api_key_authorizer_test.integration.infrastructure.api_key_secret_auth_endpoint_stack import ApiKeySecretAuthEndpointStack
+from b_cfn_custom_api_key_authorizer_test.integration.infrastructure.authorized_api_stack import AuthorizedApiStack
+from b_cfn_custom_api_key_authorizer_test.integration.infrastructure.basic_auth_endpoint_stack import BasicAuthEndpointStack
 
 
 class MainStack(TestingStack):
-    API_ENDPOINT = 'ApiEndpoint'
+    # Used to test apikey - apisecret headers.
+    API_ENDPOINT1 = 'ApiEndpoint1'
+    # Used to test authorization header.
+    API_ENDPOINT2 = 'ApiEndpoint2'
+
     AUTHORIZER_FUNCTION = 'AuthorizerFunction'
     DELETER_FUNCTION = 'DeleterFunction'
     EXISTS_FUNCTION = 'ExistsFunction'
@@ -17,45 +21,31 @@ class MainStack(TestingStack):
     def __init__(self, scope: Construct) -> None:
         super().__init__(scope=scope)
 
-        prefix = TestingStack.global_prefix()
+        self.api_stack = AuthorizedApiStack(self)
+        self.api = self.api_stack.api
+        self.stage = self.api_stack.stage
 
-        self.api = CfnApi(
-            scope=self,
-            id='Api',
-            name=f'{prefix}Api',
-            description='Sample description.',
-            protocol_type='HTTP',
-            cors_configuration=CfnApi.CorsProperty(
-                allow_methods=['GET', 'PUT', 'POST', 'OPTIONS', 'DELETE'],
-                allow_origins=['*'],
-                allow_headers=[
-                    'Content-Type',
-                    'Authorization'
-                ],
-                max_age=300
-            )
-        )
+        """
+        Endpoint 1.
+        """
 
-        self.authorizer = ApiKeyCustomAuthorizer(
-            scope=self,
-            resource_name_prefix=prefix,
-            api=self.api,
-        )
+        endpoint1 = ApiKeySecretAuthEndpointStack(self, self.api)
+        endpoint1_path = f'{self.api.attr_api_endpoint}/{self.stage.stage_name}{endpoint1.path}'
+        self.add_output(self.API_ENDPOINT1, value=endpoint1_path)
 
-        self.stage: CfnStage = CfnStage(
-            scope=self,
-            id='Stage',
-            stage_name='test',
-            api_id=self.api.ref,
-            auto_deploy=True,
-        )
+        a1 = endpoint1.authorizer_api_key_and_secret
+        self.add_output(self.AUTHORIZER_FUNCTION, value=a1.authorizer_function.function_name)
+        self.add_output(self.DELETER_FUNCTION, value=a1.deleter_function.function_name)
+        self.add_output(self.EXISTS_FUNCTION, value=a1.exists_function.function_name)
+        self.add_output(self.GENERATOR_FUNCTION, value=a1.generator_function.function_name)
+        self.add_output(self.VALIDATOR_FUNCTION, value=a1.validator_function.function_name)
 
-        self.endpoint_stack = authorized_endpoint_stack.AuthorizedEndpointStack(self, self.api, self.authorizer)
-        self.dummy_endpoint = f'{self.api.attr_api_endpoint}/{self.stage.stage_name}/{self.endpoint_stack.path}'
+        """
+        Endpoint 2.
+        """
 
-        self.add_output(self.API_ENDPOINT, value=self.dummy_endpoint)
-        self.add_output(self.AUTHORIZER_FUNCTION, value=self.authorizer.authorizer_function.function_name)
-        self.add_output(self.DELETER_FUNCTION, value=self.authorizer.deleter_function.function_name)
-        self.add_output(self.EXISTS_FUNCTION, value=self.authorizer.exists_function.function_name)
-        self.add_output(self.GENERATOR_FUNCTION, value=self.authorizer.generator_function.function_name)
-        self.add_output(self.VALIDATOR_FUNCTION, value=self.authorizer.validator_function.function_name)
+        # Allow both authorizers to share the same api keys database, because our generator functions and
+        # fixtures are looking only to a single database.
+        endpoint2 = BasicAuthEndpointStack(self, self.api, endpoint1.authorizer_api_key_and_secret.api_keys_database)
+        endpoint2_path = f'{self.api.attr_api_endpoint}/{self.stage.stage_name}{endpoint2.path}'
+        self.add_output(self.API_ENDPOINT2, value=endpoint2_path)