+++ b_cfn_custom_api_key_authorizer_test/integration/tests/test_authorizer.py
@@ -14,7 +14,7 @@ def test_FUNCTION_authorizer_WITH_valid_credentials_EXPECT_request_allowed_to_pa
 
     response = urllib3.PoolManager().request(
         method='GET',
-        url=MainStack.get_output(MainStack.API_ENDPOINT),
+        url=MainStack.get_output(MainStack.API_ENDPOINT1),
         headers={
             'ApiKey': api_key,
             'ApiSecret': api_secret
@@ -23,6 +23,30 @@ def test_FUNCTION_authorizer_WITH_valid_credentials_EXPECT_request_allowed_to_pa
 
     # Make sure response is successful.
     assert response.status == 200
+
+    data = response.data
+    data = data.decode()
+
+    # Response from a dummy lambda function defined in the infrastructure main stack.
+    assert data == 'Hello World!'
+
+
+def test_FUNCTION_authorizer_WITH_valid_basic_auth_EXPECT_request_allowed_to_pass(api_keys) -> None:
+    """
+    Tests whether the authorizer allows the request to pass through, if the valid basic auth is given.
+
+    :return: No return.
+    """
+    api_key, api_secret = api_keys
+
+    response = urllib3.PoolManager().request(
+        method='GET',
+        url=MainStack.get_output(MainStack.API_ENDPOINT2),
+        headers=urllib3.make_headers(basic_auth=f'{api_key}:{api_secret}')
+    )
+
+    # Make sure response is successful.
+    assert response.status == 200, str(response.data)
 
     data = response.data
     data = data.decode()
@@ -40,7 +64,7 @@ def test_authorizer_with_no_key_secret() -> None:
     """
     response = urllib3.PoolManager().request(
         method='GET',
-        url=MainStack.get_output(MainStack.API_ENDPOINT),
+        url=MainStack.get_output(MainStack.API_ENDPOINT1),
         headers={},
     )
 
@@ -56,7 +80,7 @@ def test_authorizer_with_non_existent_api_key_secret() -> None:
     """
     response = urllib3.PoolManager().request(
         method='GET',
-        url=MainStack.get_output(MainStack.API_ENDPOINT),
+        url=MainStack.get_output(MainStack.API_ENDPOINT1),
         headers={
             'ApiKey': '123',
             'ApiSecret': '123'
@@ -77,7 +101,7 @@ def test_authorizer_with_invalid_key_secret(api_keys) -> None:
 
     response = urllib3.PoolManager().request(
         method='GET',
-        url=MainStack.get_output(MainStack.API_ENDPOINT),
+        url=MainStack.get_output(MainStack.API_ENDPOINT1),
         headers={
             'ApiKey': api_key,
             'ApiSecret': '123'
@@ -88,7 +112,7 @@ def test_authorizer_with_invalid_key_secret(api_keys) -> None:
 
     response = urllib3.PoolManager().request(
         method='GET',
-        url=MainStack.get_output(MainStack.API_ENDPOINT),
+        url=MainStack.get_output(MainStack.API_ENDPOINT1),
         headers={
             'ApiKey': '123',
             'ApiSecret': api_secret